advantage2.vim: Add USB HID code support
In the layout file of the Advantage2 it's possible to use USB HID scan codes
instead of location tokens in remappings. An example from the manual:

[enter]>[088]

where 088 is the decimal code for Keypad Enter. The code can also be written as
88.

All the non-reserved codes in the USB HID Keyboard table are recognized, but
this does not mean that the keyboard can output them. The keyboard's manual is
a bit vague on this point.